Comp130Kivy
First changes towards completion.
Comments have been added,
Trying to implement Milly's shark mechanic but from the left to the
right so the player needs to dodge these also. Cannot seem to get it to
work at the moment.
Also trying to add a game over button to retry but isn't appearing.
43-49 is declaring the bee class.
51-55 is declaring the end game button class.
69-77 is adding the bees.
81-89 is trying to make the end game button appear but has an error when
you lose to do with "Width"
95-100 Trying to get the bees to spawn on a random Y axis(could not
think of an appropriate variable name at the time)